backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
supabase==2.5.1
langgraph==0.1.13
langchain-openai==0.1.16
langchain-core==0.2.22
pdfplumber==0.11.2
PyJWT==2.8.0
